+++ example/F0.Mvvm.Example/F0.Mvvm.Example.Windows.Input/App.xaml.cs
@@ -1,0 +1,8 @@
+﻿using System.Windows;
+
+namespace F0.Mvvm.Example.Windows.Input
+{
+	internal partial class App : Application
+	{
+	}
+}