+++ example/F0.Mvvm.Example/F0.Mvvm.Example.Windows.Input/MainWindow.xaml.cs
@@ -1,0 +1,12 @@
+﻿using System.Windows;
+
+namespace F0.Mvvm.Example.Windows.Input
+{
+	internal partial class MainWindow : Window
+	{
+		public MainWindow()
+		{
+			InitializeComponent();
+		}
+	}
+}