+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 CHANGELOG
 
 ## vNext
+- Added implementation of `Windows.Input.IInputCommand`, a parameterless `System.Windows.Input.ICommand` abstraction.
+- Added implementation of `Windows.Input.IInputCommand<T>`, a strongly typed `System.Windows.Input.ICommand` abstraction.
 
 ## v0.1.0 (2018-09-18)
-- Added ComponentModel.ViewModel implementing Property Change Notification.
+- Added `ComponentModel.ViewModel` implementing Property Change Notification.